+++ apps/nextjs/src/app/business-sign-in/page.js
@@ -1,10 +1,13 @@
 import React from 'react';
 import SignIn from './sign-in';
+import { SideNav } from '@repo/ui/side-nav';
 
 export default async function Page() {
   return (
     <div>
+      
       <SignIn />
+      
     </div>
   );
 }
