+++ apps/nextjs/src/app/business-sign-in/sign-in.js
@@ -4,7 +4,8 @@ import Link from 'next/link';
 import React from 'react';
 import { useState } from 'react';
 import { useRouter } from 'next/navigation';
-import CookieComponent from '/home/garrett/LargeProject/LargeProject/apps/nextjs/src/app/dashboard/components/CookieComponent.jsx';
+import CookieComponent from '../dashboard/components/CookieComponent.jsx';
+
 
 export default function SignUp() {
   const [businessId, setBusinessId] = useState('');
