+++ apps/nextjs/src/app/dashboard/page.js
@@ -1,6 +1,10 @@
 import { Dashboard } from './dashboard';
 import React from 'react';
+import { SideNav } from '@repo/ui/side-nav';
 
 export default async function Page() {
-  return <Dashboard />;
+  return (  
+  <SideNav />
+  
+  );
 }
